feat(搜索系统): 实现混合搜索模式和鼠标目标检测
添加鼠标目标检测器和混合搜索管理器，支持在自动搜索和鼠标指向模式间切换
新增配置选项控制鼠标目标搜索的启用和切换延迟
优化搜索时间检查逻辑，使用double类型避免精度损失

+++ src/main/java/org/yanbwe/searchcarefully/client/ClientOverlayRenderer.java
@@ -70,8 +70,8 @@ public class ClientOverlayRenderer {
                     if (hoveredSlot.hasItem()) {
                         ItemStack stack = hoveredSlot.getItem();
                         if (ItemStackHelper.hasRemainingSearchTime(stack)) {
-                            int searchTime = (int) ItemStackHelper.getRemainingSearchTime(stack);
-                            if (searchTime > 0) {
+                            double searchTime = ItemStackHelper.getRemainingSearchTime(stack);
+                            if (searchTime > 0.0) {
                                 // 只取消可取消的事件类型（如RenderTooltipEvent.Pre）
                                 if (event.isCancelable()) {
                                     event.setCanceled(true);
@@ -116,8 +116,8 @@ public class ClientOverlayRenderer {
                             if (slot.hasItem()) {
                                 ItemStack stack = slot.getItem();
                                 if (ItemStackHelper.hasRemainingSearchTime(stack)) {
-                                    int searchTime = (int) ItemStackHelper.getRemainingSearchTime(stack);
-                                    if (searchTime > 0) {
+                                    double searchTime = ItemStackHelper.getRemainingSearchTime(stack);
+                                    if (searchTime > 0.0) {
                                         activeSearchSlots.add(slot);
                                     }
                                 }
@@ -130,8 +130,8 @@ public class ClientOverlayRenderer {
                         if (slot.hasItem()) {
                             ItemStack stack = slot.getItem();
                             if (ItemStackHelper.hasRemainingSearchTime(stack)) {
-                                int searchTime = (int) ItemStackHelper.getRemainingSearchTime(stack);
-                                if (searchTime > 0) {
+                                double searchTime = ItemStackHelper.getRemainingSearchTime(stack);
+                                if (searchTime > 0.0) {
                                     activeSearchSlots.add(slot);
                                 }
                             }
