+++ src/main/java/org/yanbwe/searchcarefully/mixin/SlotRenderMixin.java
@@ -23,9 +23,9 @@ public class SlotRenderMixin {
         // 检查槽位是否有物品且需要搜索
         ItemStack itemStack = slot.getItem();
         if (!itemStack.isEmpty() && ItemStackHelper.hasRemainingSearchTime(itemStack)) {
-            int searchTime = (int) ItemStackHelper.getRemainingSearchTime(itemStack);
+            double searchTime = ItemStackHelper.getRemainingSearchTime(itemStack);
             
-            if (searchTime > 0) {
+            if (searchTime > 0.0) {
                 // 将需要渲染遮罩的槽位信息添加到缓存中
                 SlotRenderCache.addSlotOverlay((AbstractContainerScreen<?>) (Object) this, slot);
             }
